add tests for shallow validation
