+++ cloudslang-compiler/src/test/java/io/cloudslang/lang/compiler/CompilerErrorsTest.java
@@ -11,13 +11,10 @@
 package io.cloudslang.lang.compiler;
 
 import io.cloudslang.lang.compiler.configuration.SlangCompilerSpringConfig;
-import io.cloudslang.lang.compiler.modeller.model.Executable;
 import io.cloudslang.lang.compiler.parser.utils.ParserExceptionHandler;
 import java.net.URI;
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
-import org.junit.Assert;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.ExpectedException;
@@ -557,68 +554,6 @@ public class CompilerErrorsTest {
                 "\n" +
                 "Make sure a value is specified or that indentation is properly done.");
         compiler.compile(SlangSource.fromFile(resource), path);
-    }
-
-    @Test
-    public void testValidationOfFlowWithMissingNavigationFromOperationResult() throws Exception {
-        URI flowUri = getClass().getResource("/corrupted/step_with_missing_navigation_from_operation_result_flow.sl").toURI();
-        Executable flowModel = compiler.preCompile(SlangSource.fromFile(flowUri));
-
-        URI operationUri = getClass().getResource("/java_op.sl").toURI();
-        Executable operationModel = compiler.preCompile(SlangSource.fromFile(operationUri));
-        Set<Executable> dependencies = new HashSet<>();
-        dependencies.add(operationModel);
-
-        exception.expect(RuntimeException.class);
-        exception.expectMessage("Cannot compile flow 'step_with_missing_navigation_from_operation_result_flow' " +
-                "since for step 'step1' the results [FAILURE] of its dependency 'user.ops.java_op' " +
-                "have no matching navigation.");
-        List<RuntimeException> errors = compiler.validateSlangModelWithDirectDependencies(flowModel, dependencies);
-        Assert.assertEquals(1, errors.size());
-        throw errors.get(0);
-    }
-
-    @Test
-    public void testValidationOfFlowWithMissingDependencyRequiredInputInStep() throws Exception {
-        URI flowUri = getClass().getResource("/corrupted/flow_missing_dependency_required_input_in_step.sl").toURI();
-        Executable flowModel = compiler.preCompile(SlangSource.fromFile(flowUri));
-
-        URI operation1Uri = getClass().getResource("/test_op.sl").toURI();
-        Executable operation1Model = compiler.preCompile(SlangSource.fromFile(operation1Uri));
-        URI operation2Uri = getClass().getResource("/check_op.sl").toURI();
-        Executable operation2Model = compiler.preCompile(SlangSource.fromFile(operation2Uri));
-        Set<Executable> dependencies = new HashSet<>();
-        dependencies.add(operation1Model);
-        dependencies.add(operation2Model);
-
-        exception.expect(IllegalArgumentException.class);
-        exception.expectMessage("Cannot compile flow 'io.cloudslang.flow_missing_dependency_required_input_in_step'. " +
-                "Step 'explicit_alias' does not declare all the mandatory inputs of its reference. " +
-                "The following inputs of 'user.ops.test_op' are not private, required and with no default value: alla.");
-        List<RuntimeException> errors = compiler.validateSlangModelWithDirectDependencies(flowModel, dependencies);
-        Assert.assertEquals(1, errors.size());
-        throw errors.get(0);
-    }
-
-    @Test
-    public void testValidationOfFlowInputInStepWithSameNameAsDependencyOutput() throws Exception {
-        URI flowUri = getClass().getResource("/corrupted/flow_input_in_step_same_name_as_dependency_output.sl").toURI();
-        Executable flowModel = compiler.preCompile(SlangSource.fromFile(flowUri));
-
-        URI operation1Uri = getClass().getResource("/test_op.sl").toURI();
-        Executable operation1Model = compiler.preCompile(SlangSource.fromFile(operation1Uri));
-        URI operation2Uri = getClass().getResource("/check_op.sl").toURI();
-        Executable operation2Model = compiler.preCompile(SlangSource.fromFile(operation2Uri));
-        Set<Executable> dependencies = new HashSet<>();
-        dependencies.add(operation1Model);
-        dependencies.add(operation2Model);
-
-        exception.expect(IllegalArgumentException.class);
-        exception.expectMessage("Cannot compile flow 'io.cloudslang.flow_input_in_step_same_name_as_dependency_output'. " +
-                "Step 'explicit_alias' has input 'balla' with the same name as the one of the outputs of 'user.ops.test_op'.");
-        List<RuntimeException> errors = compiler.validateSlangModelWithDirectDependencies(flowModel, dependencies);
-        Assert.assertEquals(1, errors.size());
-        throw errors.get(0);
     }
 
     @Test
