+++ cloudslang-compiler/src/test/resources/shallow_validation/test_op.sl
@@ -1,0 +1,20 @@
+#   (c) Copyright 2014 Hewlett-Packard Development Company, L.P.
+#   All rights reserved. This program and the accompanying materials
+#   are made available under the terms of the Apache License v2.0 which accompany this distribution.
+#
+#   The Apache License is available at
+#   http://www.apache.org/licenses/LICENSE-2.0
+
+namespace: user.ops
+
+operation:
+  name: test_op
+  inputs:
+    - alla
+  python_action:
+    script: 'print "hello world"'
+  outputs:
+    - balla: 'some value'
+  results:
+    - SUCCESS: ${1 == 1}
+    - FAILURE
